# docs/api/controller.openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BlindPass Local Controller API",
    "version": "0.1.0",
    "description": "P03 fleet controller contract, including reviewed node, workload, policy, approval, grant, operation and node-channel schemas. Paths marked x-blindpass-status=planned are not mounted yet. This file uses JSON-compatible YAML 1.2 syntax so the checked-in type generator can use Node's standard library without installing a YAML parser.",
    "x-blindpass-ct19": "adopted",
    "x-blindpass-legacy-route-count": 12
  },
  "servers": [
    { "url": "http://127.0.0.1:3200", "description": "Local controller" }
  ],
...
  "paths": {
...
    "/api/v3/nodes/{node_id}/source-bindings/{resource_id}": {
      "get": {
        "operationId": "getFleetSourceBinding",
        "tags": [
          "Fleet administration"
        ],
        "description": "An authenticated administrator cookie is required. The binding is independently configured for one enrolled node/resource. A change invalidates later use of previously recorded offers.",
        "security": [
          {
            "adminSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "resource_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Independent source destination metadata; no Source or input URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceBinding"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AdminError"
          },
          "403": {
            "$ref": "#/components/responses/AdminError"
          },
          "404": {
            "$ref": "#/components/responses/AdminError"
          },
          "503": {
            "$ref": "#/components/responses/AdminError"
          }
        }
      },
      "put": {
        "operationId": "updateFleetSourceBinding",
        "tags": [
          "Fleet administration"
        ],
        "description": "An authenticated administrator cookie is required. The binding is independently configured for one enrolled node/resource. A change invalidates later use of previously recorded offers. Origin/CSRF and the expected version are checked; version0 creates and current positive versions update. Inactive or pending rotation/revocation nodes and stale versions return409.",
        "security": [
          {
            "adminSession": [],
            "csrfCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "resource_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "responses": {
          "200": {
            "description": "Independent source destination metadata; no Source or input URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceBinding"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AdminError"
          },
          "401": {
            "$ref": "#/components/responses/AdminError"
          },
          "403": {
            "$ref": "#/components/responses/AdminError"
          },
          "409": {
            "$ref": "#/components/responses/AdminError"
          },
          "422": {
            "$ref": "#/components/responses/AdminError"
          },
          "503": {
            "$ref": "#/components/responses/AdminError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceBindingInput"
              }
            }
          }
        }
      }
    },
    "/api/v3/admin/operations/{id}/provisioning-link": {
      "post": {
        "operationId": "createFleetProvisioningLink",
        "tags": [
          "Fleet administration"
        ],
        "description": "Creates the one immutable Source link for an original browser_session grant. The deciding operator owns Source collection for an approved grant; for an automatically allowed grant the administrator who last configured the node/resource destination owns it. A viewer, another administrator or operator, a node bearer or a desktop token is denied (403/401). Origin, CSRF and a 16-128 character Idempotency-Key are required. The transaction re-checks session, role, grant, operation, workload registration, policy, issuer epoch, enrolled key and destination version. The link expires with the original recipient offer. An exact retry (same key and operator) returns the original link with 200; any other key, or a second link for the grant, conflicts. 409 provisioning_offer_pending means the node has not yet published its offer; retry shortly. Capabilities are derived from the root secret in a fleet-specific signing domain; a legacy browser capability is never accepted.",
        "security": [
          {
            "adminSession": [],
            "csrfCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Controller operation identifier of a browser_session operation"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "responses": {
          "200": {
            "description": "Exact retry: the original immutable link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningLink"
                }
              }
            }
          },
          "201": {
            "description": "Link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningLink"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "401": {
            "description": "Session required or ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "403": {
            "description": "Role, CSRF/Origin or named-owner check failed (role_denied, csrf_or_origin_denied, provisioning_owner_required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "404": {
            "description": "operation_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "409": {
            "description": "provisioning_offer_pending or provisioning_link_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "410": {
            "description": "provisioning_unavailable: authority, offer or deadline no longer holds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "503": {
            "description": "Controller state unavailable; no link was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/fleet/provisioning/{id}/metadata": {
      "get": {
        "operationId": "getFleetProvisioningMetadata",
        "tags": [
          "Fleet administration"
        ],
        "description": "Read-only status for the link's named operator. Requires that operator's current browser session cookie and the fleet metadata capability (sig); the submit capability, a legacy browser capability, another operator or a node bearer is denied. Repeated reads and browser prefetch never consume the recipient key, create a delivery or write audit. A ready response carries the original signed recipient offer and the independently trusted expected values (the original controller-signed grant body, the enrolled broker signing key, node key version and the administrator-configured destination) in exactly the shape the browser library's verifyBrowserRecipientOffer accepts as `expected`; none of them is derived from the offer. After a successful submit the answer is the immutable receipt, even if the grant was later consumed.",
        "security": [
          {
            "adminSession": [],
            "fleetProvisioningCapability": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{64}$"
            }
          },
          {
            "in": "query",
            "name": "sig",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fleet metadata capability returned as metadata_sig"
          }
        ],
        "responses": {
          "200": {
            "description": "Ready for Source or already submitted",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProvisioningMetadata"
                    },
                    {
                      "$ref": "#/components/schemas/ProvisioningReceipt"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Session required or ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "403": {
            "description": "Missing/invalid capability, role, or not the link's named operator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "410": {
            "description": "provisioning_unavailable: expired, withdrawn or changed authority",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          },
          "503": {
            "description": "Controller state unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "agentBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "nodeBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Short-lived 15-minute node session bound to an approved node key and audience blindpass-node" },
      "externalWorkloadBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Configured external issuer with pinned issuer, audience and JWKS validation" },
      "agentApiKey": { "type": "apiKey", "in": "header", "name": "x-agent-api-key" },
      "browserSignedLink": { "type": "apiKey", "in": "query", "name": "sig", "description": "Expiry-bound HMAC signature scoped to the route operation" },
      "bootstrapCapability": { "type": "apiKey", "in": "header", "name": "x-blindpass-bootstrap-token" },
      "adminSession": { "type": "apiKey", "in": "cookie", "name": "bp_session", "description": "Local operator session cookie. While the operator must change a temporary password (bootstrap, local reset or test fixture), every administrative route other than session read, refresh, logout and change-password answers 403 with error password_change_required." },
      "adminRefresh": { "type": "apiKey", "in": "cookie", "name": "bp_refresh" },
      "csrfCookie": { "type": "apiKey", "in": "cookie", "name": "bp_csrf" },
      "desktopBearer": { "type": "http", "scheme": "bearer", "description": "P04-D3 desktop approval-app access token from a kind desktop login or refresh, sent without Origin. Accepted only by session read, refresh and logout and the /api/v3/approvals routes; every other route answers 401. It expires 20 minutes after the login or refresh that issued it, as well as on logout, idle timeout, password change, reset, removal and family revocation. Unsafe methods need no CSRF value because nothing is ambient." },
      "fleetProvisioningCapability": { "type": "apiKey", "in": "query", "name": "sig", "description": "Fleet Source link capability: expiry and a scope-bound (metadata or submit) HMAC in a fleet-specific derived signing domain, distinct from browserSignedLink. It never authorizes by itself: the link's named operator session, role and current authority are also required on every call." },
      "seedToken": { "type": "apiKey", "in": "header", "name": "x-blindpass-seed-token" }
    },
    "parameters": {
      "Origin": { "name": "Origin", "in": "header", "required": true, "schema": { "type": "string", "format": "uri" } },
...
      "CsrfHeader": { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" }, "description": "Must match the authenticated session's server-side CSRF value and bp_csrf cookie." },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 128 } },
...
    },
    "responses": {
...
      "AdminError": { "description": "Local administration error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminError" } } } },
...
    },
    "schemas": {
...
      "AdminError": { "type": "object", "properties": { "error": { "type": "string" }, "message": { "type": "string" } }, "required": ["error"], "additionalProperties": false },
...
      "SignedDocument": { "type": "object", "properties": { "kind": { "type": "string", "enum": ["registration", "policy_snapshot", "grant", "revocation", "node_revocation", "node_key_rotation", "time_reply", "application_ack", "operation_result", "operation_closed", "recipient_offer", "provisioning_delivery", "audit_event"] }, "v": { "type": "integer", "const": 1 }, "body": { "type": "object", "additionalProperties": true }, "sig": { "type": "string", "pattern": "^[A-Za-z0-9_-]{86}$" }, "kid": { "type": "string" }, "epoch": { "type": "integer", "minimum": 1 } }, "required": ["kind", "v", "body", "sig", "kid", "epoch"], "additionalProperties": false },
...
      "SourceBindingInput": {"type": "object", "properties": {"source_unit": {"type": "string", "minLength": 1, "maxLength": 255, "pattern": "^(?!\\.)[A-Za-z0-9_.@:-]+\\.service$"}, "credential": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^(?!\\.)[A-Za-z0-9_.-]+$"}, "expected_version": {"type": "integer", "minimum": 0}}, "required": ["source_unit", "credential", "expected_version"], "additionalProperties": false},
      "SourceBinding": {"type": "object", "properties": {"node_id": {"type": "string"}, "resource_id": {"type": "string"}, "source_unit": {"type": "string"}, "credential": {"type": "string"}, "version": {"type": "integer", "minimum": 1}, "updated_at": {"type": "integer"}}, "required": ["node_id", "resource_id", "source_unit", "credential", "version", "updated_at"], "additionalProperties": false},
      "ProvisioningLink": {"type": "object", "properties": {"id": {"type": "string", "pattern": "^[a-f0-9]{64}$"}, "metadata_sig": {"type": "string"}, "submit_sig": {"type": "string"}, "operator_id": {"type": "string"}, "operation_id": {"type": "string"}, "expires_at_ms": {"type": "integer", "minimum": 1, "description": "The original recipient offer deadline"}, "input_path": {"type": "string", "description": "Same-origin path that opens the ordinary input page in fleet mode: /?kind=fleet&id=...&metadata_sig=...&submit_sig=..."}}, "required": ["id", "metadata_sig", "submit_sig", "operator_id", "operation_id", "expires_at_ms", "input_path"], "additionalProperties": false},
      "ProvisioningMetadata": {"type": "object", "properties": {"status": {"type": "string", "const": "ready"}, "server_time_ms": {"type": "integer"}, "expires_at_ms": {"type": "integer"}, "offer": {"$ref": "#/components/schemas/SignedDocument"}, "expected": {"$ref": "#/components/schemas/ProvisioningExpected"}, "summary": {"type": "object", "properties": {"purpose": {"type": "string"}, "workload_name": {"type": "string"}}, "required": ["purpose", "workload_name"], "additionalProperties": false}}, "required": ["status", "server_time_ms", "expires_at_ms", "offer", "expected", "summary"], "additionalProperties": false},
...
      "ProvisioningReceipt": {"type": "object", "properties": {"status": {"type": "string", "const": "submitted"}, "offer_id": {"type": "string"}, "ciphertext_digest": {"type": "string", "pattern": "^[a-f0-9]{64}$"}, "delivery_digest": {"type": "string", "pattern": "^[a-f0-9]{64}$"}, "submitted_at_ms": {"type": "integer"}, "expires_at_ms": {"type": "integer"}}, "required": ["status", "offer_id", "ciphertext_digest", "delivery_digest", "submitted_at_ms", "expires_at_ms"], "additionalProperties": false},
...
    }
  }
}